issue #41: do a better job handling authentication failures
This defines a new broadcast action to send to the NowPlayingActivity.  Due to
the short-livedness of the service it should be the activity's job to notify
the user that reauthentication is required.  If an authentication error occurs,
the session is put into an anonymous state and the user may attempt to use the
app anonymously.