UselessParent lint roundup
Note for election_drawer.xml: the ignore here is because the parent has some
role in helping float the SlidingDrawer to the bottom of the screen though I do
not know the exact reason why anymore (it was written long ago).

In any case, the SlidingDrawer is deprecated and should be removed (issue #32).